requirements.txt
streamlit
pillow
tensorflow
numpy
openai
scikit-learn
matplotlib

